fix: ProductDetail 화면 구성 변경
